Separated The Main page from Loginpage.Dart
- Separated Main page from Loginpage.Dart

-Added a Temporary  Back bottom in MainPage.Dart for easy navigate

- Added some Comments in main.dart
