+++ safefoodie_login_merge/lib/LoginPage.dart
@@ -3,6 +3,13 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
+Future<FirebaseApp> _initializeFirebase() async {
+  FirebaseApp firebaseApp = await Firebase.initializeApp();
+  return firebaseApp;
+}
+
+final FirebaseAuth _auth = FirebaseAuth.instance;
+
 class LoginScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -90,7 +97,7 @@ class LoginScreen extends StatelessWidget {
             child: ElevatedButton(
               child: const Text('Login'),
               onPressed: () {
-                _ToMainScreen(context);
+                Navigator.pushNamed(context, 'Mainpage');
                 //print(nameController.text);
                 //print(passwordController.text);
               },
@@ -118,126 +125,10 @@ class LoginScreen extends StatelessWidget {
       ])),
     );
   }
-
-//======================================================================================
-//NAV FUNCTIONS
-//======================================================================================
-//Nav function for signup
-
-//Nav function for main page
-  void _ToMainScreen(BuildContext context) {
-    Navigator.of(context)
-        .push(MaterialPageRoute(builder: (context) => MainPage()));
-  }
-
-//Nav function for search lists page
-
-//Nav function for create list/add item search page
-
-//Nav function for settings
-
 }
-
 //======================================================================================
 //======================================================================================
 //Firebase Authorization
-Future<FirebaseApp> _initializeFirebase() async {
-  FirebaseApp firebaseApp = await Firebase.initializeApp();
-  return firebaseApp;
-}
-
-final FirebaseAuth _auth = FirebaseAuth.instance;
-
-//======================================================================================
-//PAGE CLASS INITIALIZATION
-//======================================================================================
-//Signup class
-
-//main page
-class MainPage extends StatefulWidget {
-  @override
-  _MainPageState createState() => _MainPageState();
-}
-
-//Search lists page
-
-//Add item/create page
-
-//settings page
-
-//======================================================================================
-//EMAIL SIGNUP FUNCTION
-//======================================================================================
-
-//======================================================================================
-//RESET PAGE START
-//======================================================================================
-
-//======================================================================================
-//HOME PAGE START
-//======================================================================================
-
-class _MainPageState extends State<MainPage> {
-//=======================================
-//Button Widgets
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-          child: ListView(children: <Widget>[
-//======================================
-//title header
-        Container(
-            alignment: Alignment.center,
-            padding: const EdgeInsets.all(10),
-            child: const Text(
-              'SafeFoodie',
-              style: TextStyle(
-                  color: Colors.white,
-                  fontWeight: FontWeight.w500,
-                  fontSize: 30),
-            )),
-
-        Icon(
-          Icons.bakery_dining_rounded,
-          color: Color.fromARGB(216, 230, 182, 53),
-        ),
-
-//======================================
-        SizedBox(
-          height: 200,
-        ),
-        Container(
-            child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            IconButton(
-                icon: const Icon(Icons.align_horizontal_left),
-                color: Colors.white, //view all lists
-                onPressed: () {
-//_ToViewListsScreen(context);
-                }),
-            IconButton(
-                icon: const Icon(Icons.search),
-                color: Colors.white, //search lists
-                onPressed: () {}),
-            IconButton(
-                icon: const Icon(Icons.add_circle),
-                color: Colors.white, //Create list, add item
-                onPressed: () {}),
-            IconButton(
-                icon: const Icon(Icons.settings),
-                color: Colors.white, //settings
-                onPressed: () {}),
-          ],
-        ))
-
-//======================================
-      ])),
-    );
-  }
-}
 
 //======================================================================================
 //VIEW LISTS PAGE START
@@ -335,6 +226,8 @@ class _ViewListsState extends State<ViewListsPage> {
           ],
         ));
   }
+
+// Need to import these after Viewlist page is Created //
 
 //======================================================================================
 //NAV FUNCTIONS
