+++ safefoodie_login_merge/lib/main.dart
@@ -2,9 +2,13 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+
+// Must Import .Dart inorder to build the widgets
 import 'LoginPage.dart';
 import 'Password.dart';
 import 'signup.dart';
+import 'Mainpage.dart';
+///////////////////////////////////////////////////
 
 void main() => runApp(MyApp());
 
@@ -19,14 +23,18 @@ class MyApp extends StatelessWidget {
             Color.fromARGB(166, 72, 168, 75), //color of background
         textTheme: Typography().white, //sets default text to white
       ),
-      initialRoute: 'LoginPage',
+      // Inital First Page so that LoginPage would load first
+      initialRoute: 'LoginScreen',
       home: LoginScreen(),
       routes: {
         // When navigating to the "homeScreen" route, build the HomeScreen widget.
         'LoginScreen': (context) => LoginScreen(),
-        // When navigating to the "secondScreen" route, build the SecondScreen widget.
+        // When navigating to the "SignupScreen" route, build the SignupScreen widget.
         'SignupScreen': (context) => SignupScreen(),
+        // When navigating to the "PasswordScreen" route, build the PasswordScreen widget.
         'PasswordScreen': (context) => PasswordScreen(),
+        // When navigating to the "Mainpage" route, build the Mainpage widget.
+        'Mainpage': (context) => Mainpage(),
       },
     );
   }
